+++ src/app/hotels/(hotels-info)/Hotels-Sliders.tsx
