+++ src/app/hotels/(hotels-info)/loading.tsx
