+++ src/app/hotels/(hotels-info)/style.module.css
