+++ src/app/hotels/(hotels-info)/template.tsx
