+++ src/style/style-components.css
@@ -29,7 +29,7 @@
     @import url('../components/Skeleton/skeleton-style.css');
     @import url('../components/ListBox/list-box-template.module.css');
     @import url('../app/(home)/(shared)/style.css');
-    @import url('../app/hotels/style.module.css');
+    @import url('../app/hotels/(hotels-info)/style.module.css');
 }
 
 @layer utilities {
